refactor(config): move Config to src/config below the worker layer
S2 commit 1/5. Pure move: server/state/config.* -> config/config.*
with include rewrites in all consumers. Config is carried wholesale by
the worker protocol (BuildParams/QueryParams), so it must sit below the
future clice-worker target. No code changes.
